+++ app/components/UI/Tokens/index.test.tsx
@@ -24,13 +24,15 @@ jest.mock('./TokensBottomSheet', () => ({
   createTokensBottomSheetNavDetails: jest.fn(() => ['BottomSheetScreen', {}]),
 }));
 
-jest.mock('../Earn/components/Musd/MusdConversionCta', () => {
+jest.mock('../Earn/components/Musd/MusdConversionAssetListCta', () => {
   const { View } = jest.requireActual('react-native');
-  const MockMusdConversionCta = () => <View testID="musd-conversion-cta" />;
+  const MusdConversionAssetListCta = () => (
+    <View testID="musd-conversion-cta" />
+  );
 
   return {
     __esModule: true,
-    default: MockMusdConversionCta,
+    default: MusdConversionAssetListCta,
   };
 });
